Rename single sample output file
