+++ README.md
@@ -142,7 +142,7 @@ var options = {
       // maxLinkDensity: 0.65
     },
     // reranker is disabled by default; enable after training weights
-    // Note: tests/single-sample-run.js auto-loads weights.json (if present) and enables the reranker
+    // Note: scripts/single-sample-run.js auto-loads weights.json (if present) and enables the reranker
     reranker: { enabled: false }
     // optional: dump top-N candidates per page for labeling
     // debugDump: { path: 'candidates_with_url.csv', topN: 5, addUrl: true }
@@ -353,7 +353,7 @@ Run quick tests and batches from this repo without writing code.
     - `npm run batch:crawl -- <urlsFile?> <outCsv?>`
   - sample:prepare: Fetch a smaller curated set (default 200) for quick sampling.
     - `npm run sample:prepare`
- - sample:single: Run a single URL parse and write JSON to `tests/results/`.
+ - sample:single: Run a single URL parse and write JSON to `scripts/results/single-sample-run-result.json`.
    - `npm run sample:single -- "https://example.com/article"`
  - sample:batch: Run the multi-URL sample with progress bar and summaries.
    - `npm run sample:batch -- <N> <concurrency> <urlsFile> <timeoutMs>`
@@ -372,10 +372,10 @@ Run quick tests and batches from this repo without writing code.
 
 ### Single URL test
 
-Writes a detailed JSON to `tests/results/`.
+Writes a detailed JSON to `scripts/results/single-sample-run-result.json`.
 
 ```bash
-TEST_TIMEOUT_MS=40000 node tests/single-sample-run.js "https://www.cnn.com/business/live-news/fox-news-dominion-trial-04-18-23/index.html"
+TEST_TIMEOUT_MS=40000 node scripts/single-sample-run.js "https://www.cnn.com/business/live-news/fox-news-dominion-trial-04-18-23/index.html"
 # or via npm script
 npm run sample:single -- "https://www.cnn.com/business/live-news/fox-news-dominion-trial-04-18-23/index.html"
 ```
@@ -383,7 +383,7 @@ npm run sample:single -- "https://www.cnn.com/business/live-news/fox-news-domini
 PowerShell:
 
 ```powershell
-$env:TEST_TIMEOUT_MS=40000; node tests/single-sample-run.js "https://www.cnn.com/business/live-news/fox-news-dominion-trial-04-18-23/index.html"
+$env:TEST_TIMEOUT_MS=40000; node scripts/single-sample-run.js "https://www.cnn.com/business/live-news/fox-news-dominion-trial-04-18-23/index.html"
 # or
 npm run sample:single -- "https://www.cnn.com/business/live-news/fox-news-dominion-trial-04-18-23/index.html"
 ```
@@ -405,13 +405,13 @@ Parameters
 
 - `800`: target number of URLs to collect into `scripts/data/urls.txt`.
 
-2) Run a batch against unique hosts with a simple progress-only view. Progress and a final summary print to the console; JSON/CSV reports are saved under `tests/results/`.
+2) Run a batch against unique hosts with a simple progress-only view. Progress and a final summary print to the console; JSON/CSV reports are saved under `scripts/results/`.
 
 Bash/Zsh:
 
 ```bash
 UNIQUE_HOSTS=1 SAMPLE_PROGRESS_ONLY=1 SAMPLE_TICK_MS=1000 \
-  node tests/batch-sample-run.js 100 8 scripts/data/urls.txt 25000
+  node scripts/batch-sample-run.js 100 8 scripts/data/urls.txt 25000
 # or via npm script (defaults shown in package.json)
 npm run sample:batch -- 100 8 scripts/data/urls.txt 25000
 ```
@@ -420,7 +420,7 @@ PowerShell:
 
 ```powershell
 $env:UNIQUE_HOSTS=1; $env:SAMPLE_PROGRESS_ONLY=1; $env:SAMPLE_TICK_MS=1000; \
-  node tests/batch-sample-run.js 100 8 scripts/data/urls.txt 25000
+  node scripts/batch-sample-run.js 100 8 scripts/data/urls.txt 25000
 # or
 npm run sample:batch -- 100 8 scripts/data/urls.txt 25000
 ```
@@ -441,7 +441,7 @@ You can train a simple logistic-regression reranker to improve candidate selecti
 
 1) Generate candidate features
 - Single URL (appends candidates):
-  - `node tests/single-sample-run.js`
+  - `node scripts/single-sample-run.js`
 - Batch (recommended):
   - `node scripts/batch-crawl.js scripts/data/urls.txt scripts/data/candidates_with_url.csv 0 200`
   - Adjust `start` and `limit` to process in slices (e.g., `200 200`, `400 200`, ...).
@@ -451,7 +451,7 @@ You can train a simple logistic-regression reranker to improve candidate selecti
   - `scripts/data/candidates_with_url.csv`: output CSV file for candidate features
   - `0`: start offset (row index) in the URLs file
   - `200`: limit (number of URLs to process in this run)
-- The project dumps candidate features with URL by default (see `tests/single-sample-run.js`):
+- The project dumps candidate features with URL by default (see `scripts/single-sample-run.js`):
   - Header: `url,xpath,css_selector,text_length,punctuation_count,link_density,paragraph_count,has_semantic_container,boilerplate_penalty,direct_paragraph_count,direct_block_count,paragraph_to_block_ratio,average_paragraph_length,dom_depth,heading_children_count,aria_role_main,aria_role_negative,aria_hidden,image_alt_ratio,image_count,training_label,default_selected`
   - Up to `topN` unique-XPath rows per page (default 5)
 
@@ -489,13 +489,13 @@ You can train a simple logistic-regression reranker to improve candidate selecti
   - `> weights.json` redirects stdout to a file (Bash/PowerShell)
 
 4) Use the weights
-- `tests/single-sample-run.js` auto-loads `weights.json` (if present) and enables the reranker:
+- `scripts/single-sample-run.js` auto-loads `weights.json` (if present) and enables the reranker:
   - `options.contentDetection.reranker = { enabled: true, weights }`
 
 Notes
 - If no reranker is configured, the detector uses heuristic scoring only.
 - You can merge CSVs from multiple runs: `npm run merge:csv` (writes `scripts/data/merged.csv`).
-- Tip: placing a `weights.json` in the project root will make `tests/single-sample-run.js` auto-enable the reranker on the next run.
+- Tip: placing a `weights.json` in the project root will make `scripts/single-sample-run.js` auto-enable the reranker on the next run.
 
 ### Crawl Tweaks (config-driven)
