+++ package.json
@@ -12,8 +12,8 @@
     "batch:crawl": "node scripts/batch-crawl.js scripts/data/urls.txt scripts/data/candidates_with_url.csv",
     "train:ranker": "node scripts/train-reranker.js",
     "sample:prepare": "node scripts/fetch-curated-urls.js 200",
-    "sample:single": "node tests/single-sample-run.js",
-    "sample:batch": "node tests/batch-sample-run.js 100 5 scripts/data/urls.txt 20000",
+    "sample:single": "node scripts/single-sample-run.js",
+    "sample:batch": "node scripts/batch-sample-run.js 100 5 scripts/data/urls.txt 20000",
     "docs": "jsdoc2md index.js controllers/*.js helpers.js > APIDOC.md"
   },
   "repository": {
